+++ src/code/model/utils/Point.java
@@ -7,9 +7,9 @@ public class Point implements Serializable
 {
     private static final long serialVersionUID = 1L;
     
-	private int x, y;
+	private double x, y;
 
-    public Point(int x, int y)
+    public Point(double x, double y)
     { this.x = x; this.y = y; }
     
     @Override
@@ -29,15 +29,15 @@ public class Point implements Serializable
   	public Point(Point source)
   	{ this(source.x, source.y); }
 
-    public void setX(int x) 
+    public void setX(double x) 
     { this.x = x; }
     
-    public void setY(int y) 
+    public void setY(double y) 
     { this.y = y; }
     
-    public int getX() 
+    public double getX() 
     { return x; }
 
-    public int getY() 
+    public double getY() 
     { return y; }
 }
